requirements.txt
dvc==3.59.1
fastapi==0.115.6
joblib==1.4.2
mlflow==3.9.0
numpy==2.0.2
pandas==2.2.3
pydantic==2.10.5
python-dotenv==1.0.1
scikit-learn==1.6.1
uvicorn[standard]==0.34.0
xgboost==2.1.3

# orchestration
prefect==3.6.15

# observability
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0

# monitoring storage + alerts
SQLAlchemy==2.0.37
psycopg2-binary==2.9.10
requests==2.32.3
